chore: project view list layout consistency
