+++ web/components/views/index.ts
@@ -3,3 +3,4 @@ export * from "./form";
 export * from "./modal";
 export * from "./view-list-item";
 export * from "./views-list";
+export * from "./view-list-item-action";
